+++ backend/lost/logic/file_man.py
@@ -36,10 +36,11 @@ def chonkyfy(fs_list, root, fs):
     body = root
     for idx in range((len(root.split('/')))):
         body, head = os.path.split(body)
-        folder_chain.insert(0,{
-            'id': os.path.join(body, head),
-            'name': head 
-        })
+        if head != '':
+            folder_chain.insert(0,{
+                'id': os.path.join(body, head),
+                'name': head 
+            })
     for el in fs_list:
         res = {
             'id':el['name'],
@@ -49,8 +50,9 @@ def chonkyfy(fs_list, root, fs):
             res['size'] = el['size']
         elif el['type'] == 'directory':
             res['isDir'] = True
-            #res['childrenCount'] = len(fs.ls(el['name']))
-            res['childrenCount'] = 0
+            # raise Exception(el['name'])
+            res['childrenCount'] = len(fs.ls(el['name']))
+            # res['childrenCount'] = 0
         else:
             raise Exception('Unknown file type')
         files.append(res)